Move utilities to oop.js submodule
