+++ lib/utilities.js
@@ -309,32 +309,6 @@ const nextEvent = (every, date = new Date()) => {
   return ms;
 };
 
-const makePrivate = (instance) => {
-  const iface = {};
-  const fields = Object.keys(instance);
-  for (const fieldName of fields) {
-    const field = instance[fieldName];
-    if (isConstant(fieldName)) {
-      iface[fieldName] = field;
-    } else if (typeof field === 'function') {
-      const bindedMethod = field.bind(instance);
-      iface[fieldName] = bindedMethod;
-      instance[fieldName] = bindedMethod;
-    }
-  }
-  return iface;
-};
-
-const protect = (allowMixins, ...namespaces) => {
-  for (const namespace of namespaces) {
-    const names = Object.keys(namespace);
-    for (const name of names) {
-      const target = namespace[name];
-      if (!allowMixins.includes(name)) Object.freeze(target);
-    }
-  }
-};
-
 const parseCookies = (cookie) => {
   const values = [];
   const items = cookie.split(';');
@@ -343,57 +317,6 @@ const parseCookies = (cookie) => {
     values.push([key.trim(), val.trim()]);
   }
   return Object.fromEntries(values);
-};
-
-const jsonParse = (buffer) => {
-  if (buffer.length === 0) return null;
-  try {
-    return JSON.parse(buffer);
-  } catch {
-    return null;
-  }
-};
-
-const flatObject = (sourceObject, fieldNames = []) => {
-  const target = {};
-  for (const [key, value] of Object.entries(sourceObject)) {
-    if (!isHashObject(value)) {
-      target[key] = value;
-      continue;
-    }
-    if (fieldNames.length > 0 && !fieldNames.includes(key)) {
-      target[key] = { ...value };
-      continue;
-    }
-    for (const [childKey, childValue] of Object.entries(value)) {
-      const combinedKey = `${key}${toUpperCamel(childKey)}`;
-      if (sourceObject[combinedKey] !== undefined) {
-        const error = `Can not combine keys: key "${combinedKey}" already exists`;
-        throw new Error(error);
-      }
-      target[combinedKey] = childValue;
-    }
-  }
-  return target;
-};
-
-const unflatObject = (sourceObject, fieldNames) => {
-  const result = {};
-  for (const [key, value] of Object.entries(sourceObject)) {
-    const prefix = fieldNames.find((name) => key.startsWith(name));
-    if (prefix) {
-      if (Object.prototype.hasOwnProperty.call(sourceObject, prefix)) {
-        throw new Error(`Can not combine keys: key "${prefix}" already exists`);
-      }
-      const newKey = key.substring(prefix.length).toLowerCase();
-      const section = result[prefix];
-      if (section) section[newKey] = value;
-      else result[prefix] = { [newKey]: value };
-      continue;
-    }
-    result[key] = value;
-  }
-  return result;
 };
 
 const isError = (err) => err?.constructor?.name?.includes('Error') || false;
@@ -430,11 +353,6 @@ module.exports = {
   parseDay,
   parseEvery,
   nextEvent,
-  makePrivate,
-  protect,
   parseCookies,
-  jsonParse,
-  flatObject,
-  unflatObject,
   isError,
 };
